Export AutoModerator from main package for cleaner imports (#10)
* feat: export AutoModerator from main package for cleaner imports

- Add AutoModerator to package __init__.py with __all__ export
- Update all imports from 'moderators.auto_model' to 'moderators'
- Update example model references to viddexa/nsfw-detector-mini

* style: update ruff configuration to fix CI linting failures

- Lower target version to py39 for broader compatibility
- Extend linter rules with F, RUF, and FA checks
- Add RUF-specific ignores for Unicode ambiguity and mutable defaults

* style: fix ruff linting issues across codebase

- Add missing newline at end of file in __init__.py
- Sort __all__ exports alphabetically in utils
- Remove unnecessary noqa comments from valid imports
- Remove redundant int() calls around round()
- Convert Turkish comments to English in tests

* build: bump version to 0.1.1

+++ src/moderators/utils/image.py
@@ -44,8 +44,8 @@ def preprocess_image_input(inputs: Any, min_side: int = 0) -> Any:
                 w, h = img.size
                 if w < min_side or h < min_side:
                     scale = max(min_side / w, min_side / h)
-                    new_w = int(round(w * scale))
-                    new_h = int(round(h * scale))
+                    new_w = round(w * scale)
+                    new_h = round(h * scale)
                     resample = getattr(getattr(Image, "Resampling", Image), "BILINEAR")
                     img = img.resize((new_w, new_h), resample)
             except Exception:
